// src/location/types/hab-family-name.types.ts
export enum HabFamilyName {
  FIRST_FAMILY_SINGLE = 'FIRST_FAMILY_SINGLE',
  SECOND_FAMILY_SINGLE = 'SECOND_FAMILY_SINGLE',
  SECOND_FAMILY_COMMUNITY = 'SECOND_FAMILY_COMMUNITY',
  THIRD_FAMILY_COMMUNITY = 'THIRD_FAMILY_COMMUNITY',
  FOURTH_FAMILY_COMMUNITY = 'FOURTH_FAMILY_COMMUNITY',
  RESIDENTIAL_ACCOMMODATION = 'RESIDENTIAL_ACCOMMODATION',
  ELDERLY_ACCOMMODATION = 'ELDERLY_ACCOMMODATION',
  RESIDENTIAL_COVERED_CAR_PARK = 'RESIDENTIAL_COVERED_CAR_PARK',
}